Handle converting pivot where clauses when modifying relationship queries.  Discussed with Dan where to put the method, decided on new Support\Services namespace, and to combine it with the existing RelationshipJoiner.
